# openapi.yaml
openapi: 3.0.3
info:
  title: Grant Application Portal API
  version: 1.0.0
  description: >
    Backend for the grant application portal: applicants browse grants, fill in
    a form generated from per-grant requirements, upload documents and submit;
    staff define grants, review applications and export the results.
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        details: {}
    User:
      type: object
      properties:
        id: { type: string, description: 64-bit id serialized as a string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [applicant, admin, reviewer] }
        organization: { type: string, nullable: true }
    AuthPayload:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        access_token: { type: string }
        refresh_token: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register an applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                organization: { type: string }
                phone: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthPayload' }
        '400':
          description: The e-mail is already registered
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Sign in and receive an access + refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthPayload' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Exchange a refresh token (single use — it is rotated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthPayload' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: The signed-in user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Auth]
      summary: Update your own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                organization: { type: string, nullable: true }
                phone: { type: string, nullable: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change your password (revokes every refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        '204': { description: Changed }
        '401': { $ref: '#/components/responses/Unauthorized' }
...